libgeda: Try to load standard configuration files.
Make calls to RC file loading functions also trigger attempts to load
the associated configuration files, on a best-effort basis.  Since
it's not a major problem if config files can't be loaded (that's what
defaults are for) ignore missing config files and just log a warning
when a load fails for other reasons.

When an RC file is being loaded, the appropriate configuration context
is made available via the (rc-config) function.