libgeda: eda_config_get_context_for_path() should always succeed.
If it's possible for the configuration system to fail to return a
context for a path then using the API can be hasslesome.  We maybe
should change how the configuration filename for bogus paths is
chosen, and probably add some warning messages, but it makes sense for
eda_config_get_context_for_path() to always return a configuration
context of some description.