gaf: added the --scale option
This option lets users to set a known scale for their titles and symbols
when they have to. It can be used, for instance, to keep the same known
scale for several schematics or ensure the conformity with standards
requiring using definite sizes for titles (e.g. Russian GOST).
The argument of the --scale option should be given as the output size
corresponding to 100 points in gschem (one default grid spacing), i.e.
100 mils by default.

Closes-bug: lp-1086522