Replace flag w_bounds_valid with TOPLEVEL reference
The bounding box stored with an object depends on the TOPLEVEL
context passed to the first call to world_get_single_object_bounds.
Calling this function with another TOPLEVEL object returns the stored
bounding box which may not be valid for that context.

This commit replaces the boolean validity flag with a reference to
the actual TOPLEVEL context for which the bounding box was calculated.
world_get_single_object_bounds checks if this matches the requested
TOPLEVEL context, and if not, recalculates the bounding box.