gschem: Add ability to interactively draw paths.
Paths have been in gschem for a long time, but there has only been the
ability to edit them (by moving control points) rather than to draw
them by defining a sequence of points.

This patch adds the state machine and rendering infrastructure
required for drawing paths.  Each point in the path is defined by
either clicking in the page (to add a cusp node) or by clicking and
dragging (to add a smooth node).  Paths are finished either by
clicking on the starting point (to close the path) or by clicking on
the last point (to finish an open path).