build-sys: Don't use Automake test driver unless actually necessary.
Most of the tests in geda-gaf actually provide their own test drivers
in the form of a shell script.  It's actually better just to run them
directly in a check-local rule than to use the Automake test driver
(which eats all the output now that parallel tests are being used).