gschem: Get menu key info by looking up action directly.
Until recently, many gschem actions have had two variants: 1) a
version for menus/toolbar and 2) a version for keybindings.  In order
to show the keybinding in the menus, it was necessary for each menu
item to specify both the action to run and the action for which to
show the keybinding.

Since menu- and keybinding-specific action variants have been
combined, there isn't any need for this behaviour any more.