gschem: Add some new action icons.
...because it's about time!

This patch adds icons for all of the "Add <object>" commands, and the
"Enter Select Mode" command.

The style is heavily based on the GNOME icon theme.  They get
installed to a gEDA-specific hicolor icon theme directory.

To regenerate the icons from the source SVG file, use `make
regen-svg-icons' in the gschem data directory.