gschem: Use action positions instead of -hotkey action variants.
Many gschem actions have "-hotkey" variants that are used for
keybindings so that they begin the editing action immediately rather
than waiting for the user to specify a starting point by clicking.

By checking for an action position, both variants can easily be
implemented in a single function.

Although this means that people can no longer hack their gschemrc
files to require the extra clicks, this is not a widely-exploited
capability.  Being able to make each action a single function makes
this change worthwhile.