gschem: Make popup menu work in a similar way to menu bar. [1220824]
Re-write the popup menu code to work the same way as the main menu bar
does (more-or-less).  This includes dispatching actions via Scheme and
rendering icons, although the popup menu doesn't list keybindings to
save screen area (this could change if people want it to).

The popup menu still operates from a static array defining the menu
items, because I didn't have time to make it load its definitions from
Scheme yet.

This fixes a crash bug where the popup menu was calling functions that
expected to be being dispatched via Scheme and didn't have the context
that they needed.

Closes-bug: lp-1220824
Reported-by: Nathan Stewart <therealnathanstewart@gmail.com>