Correct argument order to gdk_window_set_transient_for() in new dialogs
The first argument is the dialog, not the main window!

Ideally this functioanlity would be built into the GschemDialog class,
rather than replicated for each time it is used. The helper functions
in the GschemDialog class already set transient_for, but this does not
happen for construction via g_object_new() or by class inheritance.

A future possibility might be to add a "parent" property to the
GschemDialog class, so we can initialise things in the same way
however the dialog is constructed.