libgeda: Fix checks for circular hierarchy dependencies.
Gschem gives me the message "Hierarchy contains a circular dependency."
This is due to a bug in libgeda.  The function s_slib_search_dirs()
returns the first directory from the search path where a file is found
that includes the name of the searched file in its filename.  I was
trying to descend the hierarchy into a file called "biashk.sch".  There
is a file "ibiashk.sch" in a different directory earlier in the search
path, with the additional letter 'i' in the name. That directory is
returned.

Later, s_hierarchy_down_schematic_single() is convinced that the file
was found, but gets a found->pid==0, i.e., the same as the toplevel pid,
and thus thinks it's circular.

Somewhere in between is probably a failed open() that was not properly
checked.

  (stephan)blaulicht:~/solo/eda/cospi$ cat gafrc
  (component-library "../cecederena")
  (component-library "../irena")
  (component-library "../lib")
  (component-library ".")
  (source-library (getenv "PWD"))
  (source-library "../cecederena")
  (source-library "../irena")
  (source-library "../lib")
  (stephan)blaulicht:~/solo/eda/cospi$ ls -1 ../*/*biashk.sch*
  ../cecederena/biashk.sch
  ../cecederena/biashk.sch~
  ../irena/ibiashk.sch

gnetlist failed silently.

This patch fixes the problem for me.

Reviewed-by: Peter TB Brett <peter@peter-b.co.uk>