scheme-api: Add support for C closures.
Some Guile APIs require a procedure, but do not provide the ability to
pass a C function + user data pair (for example, the hooks API).

This patch adds the edascm_c_make_closure() function, which takes a C
function and data pointer and returns a Guile procedure that closes
around them and accepts any number of arguments.