scheme-api: Add EdascmHookProxy class.
It's sometimes useful to be able to hook a GObject up to receive
signals whenever a Guile hook is run.  An EdascmHookProxy object makes
this easier by taking care of all of the interfacing with Guile.
Other GObject instances can just connect to EdascmHookProxy's "run"
signal and receive a callback whenever the hook fires.

The "run" signal requires a callback of the form:

  void user_function (EdascmHookProxy *proxy, SCM args,
                      gpointer user_data);

where args contains the argument list originally passed to (run-hook).