build-sys: Require Automake 1.11 for developers.
A newer version of Automake is required in order to ensure that the
parallel test driver is available.

In newer versions of Automake, this new driver is used by default and
it's necessary to use

  AM_INIT_AUTOMAKE([serial-tests])

in order to get the (deprecated!) serial test driver currently used
throughout the geda-gaf package.  This doesn't provide a solution: at
the moment, we have the problem that 1) newer versions of Automake
choke on the test suite without the "serial-tests" option; 2) older
versions of Automake choke on the "serial-tests" option.

The only option available to retain the ability for all
currently-active developers to run the test suite is to port it to the
parallel test driver.  1.6.0 is too old; 1.11.0 was released in 2009
and seems to provide a reasonable balance between maturity, widespread
availability and features.  Note that users without Automake >= 1.11.0
will still be able to build tarballs, but will need to upgrade to
build from git.

(imported from commit 0dd51a48d23c2f45ce1040fceadf653dbba008e7)