gschem: Add GschemCellRendererIcon class.
In gschem we'd like to be able to associate a single icon name with
each action.  If the name matches a stock icon, we'd like to show it,
but if not we'd like to fall back to looking up the icon in the
current icon theme.

Unfortunately, this behaviour isn't possible when using a
GtkCellRendererPixbuf: you have to *either* use "icon-name" *or*
"stock-id".  The GschemCellRendererIcon subclasses
GtkCellRendererPixbuf and provides the desired fallback behaviour via
its "icon-id" property.

Affects-bug: lp-1227368