gschem: Fix multi-attribute editor lockup when right-clicking during edit.
This fixes a lockup when right-clicking on the text-edit box for the
attribute value column during an edit. The context menu being popped
up would cause our cell to loose focus, and the editing to be cancelled.

In general, it feels nicer to leave the edit open until the user explicitly
cancels it anyway, and that fixes this issue - whilst departing from the
standard GTK behaviour. (Which is to cancel a cell edit if the edit box
looses focus, but, WITH special case code to handle the context menu case).

A followup patch will be required to add consistent focus / edit cancelling
behaviour on the name column, which uses the standard GTK widgets for editing.