Revert "Move selection monitoring out of the multi attrib widget."
This reverts commit 1aaf6613a7c6991d858c1143ed3917675a115a5a.

This is the first step in implementing attribute editing multiple for
multiple objects at once. The SELECTION object is actually a smart list
of OBJECTs, with list-change notification. Passing this to the dialog
is NOT a break of abstraction, no more than passing a GList of OBJECTs
would be.

Think of the "selection" property of the multi-attribute editor as a
variable list of objects for it to edit, and there is no reason for
separating out this code.