Fix zooming (broken in commit fb958cc4f6aec2b2a3b13a72f78d3dff2ae6ada9)
Attach our x_event_scroll() handler to the drawing area, not the main
window, so we receive the event before the GtkScrolledWindow container.

Futhermore, x_event_scroll() now returns TRUE, which stops further
processing of the event after our handler. This avoids the built-in
GtkScrolledWindow signal handler also trying to act on the events.