gnetlist: output errors with non-zero exit status
Backends bom and bom2 output errors when there is no "attribs" file or
any options to specify an alternative file or attributes used are given.
In such a case the exit status of gnetlist must be non-zero to indicate
the error in order to use gnetlist, e.g, in Makefiles.

Signed-off-by: Edward Hennessy <ehennes@sbcglobal.net>