gschem: Change selection rules to make it easier to deselect nets.
gschem has a feature for selecting all connected net segments on a
sheet.  It works by looking for a click on a selected net segment in
select mode.  Unfortunately, this prevented deselection of a net
segment using Shift+<mouse-1> or Ctrl+<mouse-1> if that net segment
has any connected, unselected net segment.

This patch prevents selection of connected net segments if a modifier
key is pressed.

Closes-bug: lp-1227903