gnetlist: add support of output to stdout for most of backends
The "-o" switch can now be used with "-" for almost all the gnetlist backends
(excluding yet allegro, gsch2pcb, osmond, spice-sdb, and vams  for various
reasons) to output data directly to stdout instead of a file.
Some messages in gnetlist.c & gnetlist.scm were redirected to stderr in order
to not mingle them with netlist data when stdout is redirected to other files
or pipes.

Affects-bug: lp-700287