gnetlist: redirected some gnetlist messages to stderr
Before now, all backends used file ports explicitly and their output
did not mix with possible gnetlist error messages that went to stdout.
Now, as many of the backends may output their data to stdout, those
error messages may mix with the output of the backends, which may lead
to failing of some tests. So, the messages have been redirected to not
affect stdout.

Affects-bug: lp-700287