gschem: don't set the "active" property for preview in file selection dialog
The reason is the same as for changes in commit 0ceb1ea but this affects
the file selection dialog preview widget. Setting of the "active"
property while the dialog has not been yet realized will lead to several
assertions and failing of displaying previews at second and subsequent
times you open the dialog.