gschem: Introduced PATHMODE.
All path states have been replaced with that mode in combination
with the inside_action variable. Setting and resetting of this
variable have been moved into appropriate path functions.

Although gschem did not crash on Ctrl-x for paths, that was
because the code for paths did not always use corresponding
assertions for checking w_current->inside_actions as did the code
for other primitives. This patch adds both the assertions and the
code to fix such crashes for paths.

Affects-bug: lp-900122