gschem: Prevent moving of attributes after edit
This patch fixes the regression introduced in commit 80ae51c.
Edited attribute sticks to the mouse after the edit if its editing
is started with double-click. The reason of such behaviour is
emitting of the left button press event, which sets the
w_current->inside_action flag, before the double-click event. The
modal edit attribute dialog intercepts the release button event so
inside_action is still active after closing it. Further, a large
offset of the mouse pointer invokes changing of the current mode
to MOVEMODE or (in some cases) to SBOX. Manual resetting of the
inside_action flag prevents this behaviour.

Reported-by: Kai-Martin Knaak <knaak@iqo.uni-hannover.de>