gsch2pcb: Fixed incorrect setting of the pcb m4 directory.
In commit 9303b37, instead of using the PCBM4DIR variable set in the
configure script the literal string "PCBM4DIR" was used as the pcb m4
directory, which is obviously wrong. This patch adds another constant
storing the configured value of PCBM4DIR which will be used if the env
var PCBDATA is not set. If neither PCBM4DIR nor PCBDATA are set, the
fallback path "./m4" is used.

Affects-bug: lp-1463178