gschem: redesign find text functionality
The find text operation now provides a widget that lists all occurances of the
text objects that contain the search string.

When searching the hierarchy, the find text operation does not interfere with variables
related to the GUI. The find text operation does not switch the current schematic in
the view or interfere with the most recent page up.

The find text operation also searches all open pages instead of the current one.

The behavior of the find text operation works with multiple toplevel windows open while
running concurrent find text operations.