gschem: Fix g_rc_gschem_version().
Guile since version 1.9.6 does not check source properties of evaluated
files for some reason returning just #f for them. So the return value of
g_rc_rc_filename() if always #f when it checks rc_filename. Obviously,
we cannot use this value in scm_to_utf8_string(). This situation appears
only if wrong/old gschem-version value is set in a rc file. This patch
checks the returned value and sets rc_filename to "unknown" if guile
could not determine it.