gschem: Fix close-page! in GUI.
In gschem the close-page! procedure defined in the (geda page)
module was overridden in order to carry out all necessary GUI
clean up before exiting. However, this didn't work due to a typo
in its lowlevel Scheme function name, which caused segfaults by
using it, e.g., as GUI macro. Now it segfaults not so often :)