libgeda: Add eda_get_system_data_dirs()
Add a new function that returns a list of directories to search for
gEDA data that's shared between users on the system.  When searching
for system gEDA data files, the paths returned by
`eda_get_system_data_dirs()` should be searched in order.

The paths returned by `eda_get_system_data_dirs()` are determined
according to the following rules:

* If `$GEDADATA` is set, use it (deprecated).  Otherwise, use the
  "gEDA" subdirectory of each entry in `g_get_system_data_dirs()`.

* On non-Windows platforms only: if `GEDADATADIR` is defined, append
  it to that list (needed for weird `--prefix` settings).  Don't
  define `GEDADATADIR` if `--enable-relocatable`.

* If `$GEDADATA` is not set, find a directory containing
  `scheme/geda.scm`, and set `$GEDADATA` to it (deprecated).