libgeda: Add eda_get_system_config_dirs()
Add a new function that returns a list of directories to search for
gEDA configuration that's shared between users on the system (such as
the `geda-system.conf` file).  When searching for such files, the
paths returned by `eda_get_system_config_dirs()` should be searched in
order.

The paths returned by `eda_get_system_config_dirs() are determined
according to the following rules:

* If `$GEDADATARC` is set, use it (deprecated).  Otherwise, if
  `$GEDADATA` is set, use that (also deprecated).  Otherwise, use the
  "gEDA" subdirectory of each entry in `g_get_system_config_dirs()`.

* On non-Windows platforms only: if `GEDADATARC` is defined, append it
  to that list (needed for weird `--prefix` settings).  Don't define
  `GEDADATARC` if `--enable-relocatable`.