libgeda: Add eda_get_user_{data,config}_dir().
The `eda_get_user_data_dir()` and `eda_user_config_dir()` functions
return the places whence per-user data should be loaded and per-user
configuration should be stored, respectively.

For backwards compatibility (and if deprecated features are enabled),
if a `.gEDA` directory is present in the user's home directory, that
is returned by both functions.  Otherwise, they use a `gEDA`
subdirectory of the respective platform directories.