gschem: Add all system paths to GTK icon theme search path.
It's necessary to put all the "icons" subdirectories of the gEDA
system search path on the end of the GTK icon theme search path
(because we don't necessarily know how the search path is going to
relate to the install location).

This shouldn't really be needed in the first place -- gschem-specific
icons should be installed to the system's "hicolor" icon theme.