libgeda: Reimplement geda-{data,rc}-path for backwards compatibility
Up until now, the `geda-data-path` and `geda-rc-path` variables in the
rc support library have evaluated to a single directory, which has
been either from environment variables, or if that wasn't set, the
paths compiled into the gEDA tools.  This is incompatible with
supporting multiple paths to be searched for data.

For backwards compatibility, libgeda now always sets `$GEDADATA` to an
appropriate value, (unless `--disable-deprecated` was passed to
configure), so the environment variables can be relied on for setting
`geda-data-path` and `geda-rc-path`.