storage: Fix non-portable code
ISO C++ states ("forward iterator requirements") that if a and b are
forward iterator objects, they must support the expression a == b and
the result of such expression must be convertible to bool.

operator==() can be either a class member function or not.  The
current code assumed that it IS a class member, thus relying on
implementation details.