gnetlist: Set up eda_config settings once.
Apart from making things a bit faster, this also fixes a bug which may
be encountered in the current test suite where source schematics are not
placed in the same directory as their toplevel one or in its
subtree. This leads to reading of settings from different configs which
is obviously wrong for all gnetlist hierarchical settings.