libgeda: Convert always promote attribs setting to pointer array
We iterate over the always promote attributes a lot -- once for every
visible attribute found in every component that is ever loaded or
placed.

When the `TOPLEVEL::always_promote_attributes` was a `GList` of
strings, checking whether an attribute name was in the list required:

1. Parsing the attribute from the text string contents

2. Iterating over the `GList` at a cost of two pointer dereferences
   per iteration, one to get the string buffer and one to get the next
   list element

3. Performing a deep string comparison of the parsed name to the
   attribute name string by calling `strlen()`

Using interned strings allows checking for name matches by comparing
pointers, eliminating the name parsing step and the call to
`strlen()`.

`GList` is a good choice for ordered collections that are regularly
modified.  However, in this case, the list of promotable names is
almost *never* modified.  Using a `GPtrArray` allows the iteration
to take place at almost zero cost.  Usually, the number of entries in
this list is small enough that the full array of pointers will fit
into a single cache line.