ci: Provide SRFI-64 support for old Guile
Ubuntu Trusty only has Guile 2.0.9, but we need SRFI-64 test library
support, which is only available in much newer releases of Guile.

This patch adjusts the Travis CI build rules to grab the SRFI-64
library directly from the Guile git repository and put it in a
sensible place in the build worker's filesystem.