gsymcheck: add newlines to logging, fix warning pluralisation
Make the log handler in gsymcheck follow the liblepton behaviour
by adding a newline to the end of each message. Change some of the
end of check messaging to correctly build single line strings and
fix the pluralisation of the total number of warnings. Update the
tests that have a single warning to the new format.