export: Rewrite a prepared CTM to allow further modifications.
The matrix has been rewritten and supplied with comments showing
how transformations are going.  This will allow for further PS
transformations, and seems to be simpler to understand. Well, I
did so because the further transformation functions I need must be
called before the transformations described by the matrix, and I
don't know how to make a new matrix to address this.