Add support for REPL readline history in gnetlist and gschem.
New module (geda repl) has been introduced, which is used both by
gnetlist and gschem.  The same readline history file
".lepton_history" residing in the user configuration directory is
used for all tools using this module.

The REPL history is saved if the user quits the REPL explicitly,
or, in gschem, if it closes gschem. The latter is not recommended
though, since it garbles terminal settings and the user needs to
`reset` the terminal to make it functional again.

Loading the module (ice-9 session) in REPL adds support for such
procedures as `apropos' and `help'. Loading (system repl command)
adds support for guile metacommands.