gschem: Free the result of x_image_get_type_from_description() outside the function.
Internally, this function used both const and non-const char*
variables, which made it impossible to free non-const value, since
the caller didn't know if it was constant or not. Now it always
returns gchar* (constant values are copied with g_strdup()) and
the result value is g_free'd after processing in the functions
calling it. The corresponding argument in x_image_lowlevel() has
been made a constant in order to force compiler report cases of
freeing the 'filetype' variable in wrong places.