symcheck: Refactor slot/net/pinnumber duplicate checks.
Two modules have been added:
  (symbol check entity-pin)
  (symbol check net-attrib)

The modules implement slot mapping and net mapping of symbol pins,
as well as checks for duplicates in slotdef= and net= attributes,
so that check-symbol() returns the `entity-pin's of the processed
symbol.

Checks for duplicates in wrongly formed slots are avoided. Other
checks have been improved in order to make the user be able to
select any blamed attribute in lepton-schematic GUI.

Misleading warnings about matching pinnumber in net= and
pinnumber= attributes are avoided as well.