netlist: Fix `make distcheck'.
New nc symbols have inherited attribute "symbol=nc", which leads
to checking for "no connection" errors by drc2 backend. It is not
needed for the schematic
`net-not-driven-with-dont-check-if-driven-directive.sch'.  To work
around this in this case where the symbol is used for some other
goal, the symbol= attribute with other value has been attached.