netlist: Add modules '(netlist schematic toplevel)' and '(netlist deprecated)'.
(netlist schematic toplevel) contains accessors for toplevel
schematic which is now obtained as a set of schematic files given
on lepton-netlist command line.

(netlist deprecated) will contain the set of legacy (and since
that time deprecated) gnetlist procedures that should be
eventually replaced by new functions. For now, it contains an only
procedure not used in the code any more, get-pins(), and its
alias, gnetlist:get-pins(), and a set of legacy variables.