schematic: Mnemonic widgets in property tables.
Labels passed to gschem_dialog_misc_create_property_table()
can now contain mnemonic keys it their text. In
that case the widget associated with the label
(positioned next to it) will be set as a mnemonic
widget for that label. Now accelerator keys can
be used within property table widgets.