schematic: Fix doc searching bug in the `gschemdoc' module.
Sometimes, when `gschemdoc' tried to "guess" the doc path, it
failed due to using #f in constructing of resulting strings.
Now the results of guessing are filtered so this possibility is
eliminated. The bug appeared on some systems when the user chose
an entry from the Help menu.

github issue #285