netlist: Get rid of utils/.
The directory contained an only utility, mk_verilog_syms, which is
pretty much obsolete now. It was used to initially create verilog
symbols, and output symbols in obsolete format with limited set of
pin attributes. By now, initially created verilog symbols have
been modified in various ways, including manual intervention
and (IIUC) generating the symbols with some new version of the
program that did not go to the repository however. So, its removal
is safe.