netlist: Move file to page transformation functions into (netlist page) module.
The new module exports procedure filename->page which is now used
instead of file->page. In comparison with the latter, the new
procedure makes better job by returning already opened page for
given filename if it exists. An optional argument "new-page?"
allows always return a new page, which is used for backward
compatibility.