desktop: Add main category into desktop files.
According to
https://specifications.freedesktop.org/menu-spec/latest/apa.html,
previously there was no valid main category in the desktop files.

Issue description in #360:
"The problem is that without a main category, it is not possible to
know for sure if a freedesktop compliant menu will show these 2
applications."

This has been fixed by adding "Development" as main category for
Lepton. Two additional categories have been rearranged in order
"Electronics" to have priority over "Engineering".