liblepton: Fix Scheme API for arcs.
Since its creation, the function arc-end-angle() gave improper
results because it was a sibling of a C function returning the
sweep angle of an arc.  The function name and help documentation
have been fixed and now reflect the change.  For backwards
compatibility, same named function returning the proper value has
been added.