liblepton: Make bitmap-directory() return the value set.
Now, if the function bitmap-directory() is used without an
argument, it returns the value set, and it is now available for
using in Scheme scripts.

In guile-2.2.4 scm_from_utf8_string() doesn't produce zero length
strings from NULL, instead, it segfaults. To deal with the case,
an apparent check for NULL has been added.  In such a situation
the function returns #f to inform the user that the value is not
set.