schematic: Use proper exit function.
On exit, check if the GTK main loop is running
and call gtk_main_quit() or exit() depending on
that. This way we prevent errors if the program
terminates before the main loop has a chance to
start.
For example, this can happen if we want to
measure the application's startup time with
the following command:
time lepton-schematic -c '(file-quit)'