schematic: Fix visual presentation of actions in the Hotkey dialog.
Since 442e595, neither icons nor action strings were displayed in
the hotkeys dialog due to replacement of a self-evaluating
function being the core of an action with a simple one.  Initial
issue was broken action system in lepton-schematic when built with
guile-2.2.  The issue was discussed on guile-user mailing list,
and Mark H Weaver explained that inconsistent behaviour in various
guile versions and offered a solution for it:

<quote>
If you need to generate a unique _procedure_, there's only one
future-proof way to do it: the new procedure must behave differently
than every other procedure, for some input.

It's a mistake to expect procedures with equivalent behavior to be
distinguishable in Scheme.
</quote>

The solution offered and implemented is creating an action core as
case-lambda with a unique tag (list #f) that every time results in
a unique function.