netlist: Add two procedures for schematic creation.
Two new procedures, page-list->schematic() and
file-name-list->schematic() have been added to provide users with
more intuitive way of <schematic> record creation.  The previously
present procedure make-toplevel-schematic() is still there, and
has been made an alias of the latter.  The 'netlist-mode' argument
has been made optional and no longer needed to be set, which
simplifies the usage of the procedure.  By default, it is equal to
"'geda".