netlist: Add get-connection-by-netname().
The function is needed for merging connections by net name and
thus removing the need of using legacy <pin-net> record for this.
The function get-connections() has been changed to not use the
fields of this record, using the new function instead.

The modified code using this function seems to work better than
the previous one. It accounts for the output port element without
refdes in hierarchy4.sch, and processes it the same as the input
port. The tests have been updated accordingly.