netlist: Set initial hierarchy tag to '().
Since the hierarchical tag is usually a list, the empty initial
list is better to create hierarchical names than #f.

As it turns and is obvious from the updated tests, the patch also
fixes wrong using "netname=" attribute instead of "net=" when
the 'netname-priority' config setting is set to #f, that is,
"net=" should be preferred.