liblepton: Add version_message().
This function returns version message string that
can be used in the --version output of tools in
the Lepton EDA suite. All needed parameters are
substituted. The message looks like this:

Lepton EDA 1.9.9.20191003 (git: c194675)
Copyright (C) 1998-2016 gEDA developers
Copyright (C) 2017-2019 Lepton EDA developers
This is free software, and you are welcome to redistribute it
under certain conditions. For details, see the file `COPYING',
which is included in the Lepton EDA distribution.
There is NO WARRANTY, to the extent permitted by law.