upcfg: Fix call to upcfg-log().
The utility writes a message to the log with info
on which configuration is being converted.
When lepton-upcfg is invoked with some file path
rather than options, the message should include
that file path, which of course is known from the
command line parameter. So, config-file-path()
function should not be used. Otherwise, this
function will throw an exception, since it
expects to receive a configuration id,
not a string.