lepton: Add/redefine procedures in module (lepton version).
New procedures are:
- lepton-version-data() which can return the list of version
  strings by their symbol names.
- lepton-version-ref() which exactly repeats the previous
  behaviour of lepton-version() when it was invoked with one
  symbol argument.

lepton-version() has been redefined in such a way that without
arguments it returns a predefined string usually output by Lepton
tools. When it is called with arguments, the first of them must be
a format string to format output by means of simple-format, and
all others must be predefined symbols used to retrieve necessary
data the same way as for the procedures above.

Version code in the suite tools has been fixed/simplified
accordingly.