gui,gtk3: Fix segfault in gtk_tree_store_clear().
The behaviour of the functions somehow changed in GTK3 so it
may cause recursive invocation of itself on same data which leads
to segfault.

The crash can be reproduced by opening Page Manager dialog and
fiddling with pages of an hierarchical schematic (go down to
schematic or symbol, close some tabs).