attrib: Fix modification status update.
We track text value of the current cell, and if
it is modified after the cell is deactivated, the
current document is marked as dirty.
We need to handle an additional use case, when
the user, while editing text in a cell, instead
of pressing Enter to commit the changes (and,
therefore, jumping to the next cell) presses
Ctrl+S (Save).
Update the text being tracked just after that,
so that when another cell is activated,
on_deactivate() handler won't mark the document
as changed.