gui: Collapsible "Add Attribute" in multiattrib.
Replace the bottom "Add Attribute" frame in
the attributes editor dialog with a collapsible
section widget.
Most of the time this section is not used and,
on the other hand, often there's not enough room
to display all of the attributes (especially if
the "Show inherited attributes" check box is on).