lepton: Fix tests for set-object-stroke!().
The tests were broken due to new way of checks whether a page is
dirty.  Now, if a setter function does not actually modify an
object on the page by setting the same value for its property it
already has, the page is not considered as changed, there's no
point for it.

Just commented out 2 tests for geda; there's no more backward
compatibility in behaviour here.