lepton: Fix tests for set-arc!().
The tests were broken due to new way of checks whether a page is
dirty.  Now, if the setter function does not actually modify an
object on the page, that is, it sets the same value for its
property it already has, the page is not considered to be changed.

Just amended 2 'geda:' tests, they are no longer backward
compatibile though those are really of low importance.