cli: Rewrite 'lepton-cli shell' in Scheme.
The new utility is named 'lepton-shell'.

The new command does not use C function set_guile_compiled_path()
as the variable GUILE_LOAD_COMPILED_PATH is set in the script's
shebang.

The script does not check if the module (ice-9 readline) is
available on MinGW since, well, first we did not support the
platform for several years, and second, we do not know if current
MinGW supports it today.  Thus, activate-readline() is always
enabled.

lepton-cli now invokes lepton-shell if "shell" command is given.
The first "shell" argument appears when lepton-shell is invoked
within lepton-cli using execle.  It is dropped to prevent command
misinterpretation.

Snarfing code has been eliminated.