lepton: Extend geda-object->pointer*() syntax.
Now, if the syntax is used with 4 arguments, the last two are the
function checking the type of the object and the type as a symbol
or a string for prompting the user of what type is expected.  The
error message is similar to what Guile outputs in several
functions when wrong type argument is used.