gui: Fix critical glib warnings on non-text objects.
The function is_object_hit() searches for objects omitting
invisible text ones if hidden text is not shown.  The function has
been changed to check if an object is really a text object before
trying to get its 'visibility' field in order to prevent critical
warnings on hitting non-text objects.