lib: Introduce default values for LeptonFill members.
Previously, the fields of the struct were initialized to zeros.
Now, more suitable defaults are used.  They are also used in glib
check functions.  The reason for this is that while zero value is
acceptable for fill stroke width and angles, it is meaningless and
even dangerous when used for pitches.

The commit also fixes wrong type values returned by the glib check
functions.